Missed some files in the prev commit

+++ churnr/utils.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 import os
-import shutil
 import logging
 import google.cloud.storage as gcs
 
@@ -25,10 +24,6 @@ def extract_dataset_to_disk(datapath, tablenames, project, gsoutput):
 
     logger.info('Extracting {} files to {}...'.format(len(tablenames), datapath))
 
-    if os.path.exists(datapath):
-        shutil.rmtree(datapath)
-    os.makedirs(datapath)
-
     client = gcs.Client(project)
 
     split_uri = gsoutput.split('/')
@@ -37,18 +32,21 @@ def extract_dataset_to_disk(datapath, tablenames, project, gsoutput):
     bucket = gcs.Bucket(client, bucket_name)
 
     filepath = '/'.join(split_uri[3:])
-    cntr = 0
+    local_files = []
     for filename in tablenames:
-        if cntr % 10 == 0:
-            logger.info('{} out of {} files imported...'.format(cntr, len(tablenames)))
-        cntr += 1
 
         # try unsharded file first
         blob = gcs.Blob(name=os.path.join(filepath, filename), bucket=bucket)
         if blob.exists():
             local_file = os.path.join(datapath, filename)
-            with open(local_file, 'wb') as f:
-                blob.download_to_file(f)
+            if not os.path.exists(local_file):
+                logger.info('-- Downloading blob [{}] to local file [{}]'.format(blob.path, local_file))
+                with open(local_file, 'wb') as f:
+                    blob.download_to_file(f)
+            else:
+                logger.info('-- File [{}] already found on local disk, reusing it...'.format(local_file))
+
+            local_files.append(local_file)
             continue
 
         # sharded file
@@ -58,15 +56,22 @@ def extract_dataset_to_disk(datapath, tablenames, project, gsoutput):
             count = 0
             while blob.exists():
                 local_file = os.path.join(datapath, filename_shard)
-                logger.info('Downloading blob {} to local file {}'.format(blob.path, local_file))
 
-                with open(local_file, 'wb') as f:
-                    blob.download_to_file(f)
+                if not os.path.exists(local_file):
+                    logger.info('-- Downloading blob [{}] to local file [{}]'.format(blob.path, local_file))
+                    with open(local_file, 'wb') as f:
+                        blob.download_to_file(f)
+                else:
+                    logger.info('-- File [{}] already found on local disk, reusing it...'.format(local_file))
+
+                local_files.append(local_file)
                 count += 1
                 filename_shard = filename + '{0:012d}'.format(count)
                 blob = gcs.Blob(name=os.path.join(filepath, filename_shard), bucket=bucket)
         else:
             raise Exception('Blob {} not found in GCS!'.format(os.path.join(filepath, filename)))
+
+    return local_files
 
 
 def get_table_names(conf):
